+++ Sources/MessagingPush/PushHandling/PushClickHandler.swift
@@ -4,7 +4,11 @@ import UserNotifications
 
 @available(iOSApplicationExtension, unavailable)
 protocol PushClickHandler: AutoMockable {
-    func pushClicked(_ push: PushNotification)
+    // Cleanup files on device that were used when the push was displayed. Files are no longer
+    // needed now that the push is no longer shown.
+    func cleanupAfterPushInteractedWith(for push: PushNotification)
+    func trackPushMetrics(for push: PushNotification)
+    func handleDeepLink(for push: PushNotification)
 }
 
 @available(iOSApplicationExtension, unavailable)
@@ -18,19 +22,15 @@ class PushClickHandlerImpl: PushClickHandler {
         self.messagingPush = messagingPush
     }
 
-    // Note: This function is called from automatic and manual push click handlers.
-    func pushClicked(_ push: PushNotification) {
+    func trackPushMetrics(for push: PushNotification) {
         guard let cioDelivery = push.cioDelivery else {
             return
         }
 
         messagingPush.trackMetric(deliveryID: cioDelivery.id, event: .opened, deviceToken: cioDelivery.token)
+    }
 
-        // Cleanup files on device that were used when the push was displayed. Files are no longer
-        // needed now that the push is no longer shown.
-        cleanupAfterPushInteractedWith(push: push)
-
-        // Handle deep link, if there is one attached to push.
+    func handleDeepLink(for push: PushNotification) {
         if let deepLinkUrl = push.cioDeepLink?.url {
             deepLinkUtil.handleDeepLink(deepLinkUrl)
         }
@@ -39,7 +39,7 @@ class PushClickHandlerImpl: PushClickHandler {
     // There are files that are created just for displaying a rich push. After a push is interacted with, those files
     // are no longer needed.
     // This function's job is to cleanup after a push is no longer being displayed.
-    func cleanupAfterPushInteractedWith(push: PushNotification) {
+    func cleanupAfterPushInteractedWith(for push: PushNotification) {
         push.cioAttachments.forEach { attachment in
             let localFilePath = attachment.localFileUrl