+++ Sources/MessagingPush/PushHandling/iOSPushEventListener.swift
@@ -43,12 +43,26 @@ class IOSPushEventListener: PushEventHandler {
 
         logger.debug("Push came from CIO. Handle the didReceive event on behalf of the customer.")
 
+        pushClickHandler.cleanupAfterPushInteractedWith(for: push)
+
         if pushAction.didClickOnPush {
-            pushClickHandler.pushClicked(push)
+            pushClickHandler.trackPushMetrics(for: push)
         }
 
-        // call the completion handler so the customer does not need to.
-        completionHandler()
+        // Forward event to other push click handlers so they can receive a callback about this push event and optionally process the event.
+        // This funcion is an async operation that calls code we do not own. Therefore, there is risk that the completion handler will not be called and the rest of our code will not be executed. That's why it's important that before we perform this call, we do as much push processing as we can to increase SDK reliability.
+        pushEventHandlerProxy.onPushAction(pushAction, completionHandler: {
+            // When this block of code executes, the customer is done processing the push event.
+
+            // We do not open deep link until after customer is done processing the push event in case the deep link would leave the app.
+            // We want to make sure the customer has a chance to process the push event before leaving the app.
+            if pushAction.didClickOnPush {
+                self.pushClickHandler.handleDeepLink(for: push)
+            }
+
+            // call the completion handler, indicating to the OS that we are done processing the push.
+            completionHandler()
+        })
     }
 
     func shouldDisplayPushAppInForeground(_ push: PushNotification, completionHandler: @escaping (Bool) -> Void) {
@@ -75,9 +89,17 @@ class IOSPushEventListener: PushEventHandler {
 
         logger.debug("Push came from CIO. Handle the willPresent event on behalf of the customer.")
 
-        let shouldShowPush = moduleConfig.showPushAppInForeground
+        // Forward event to other push handlers so they can receive a callback about this push event.
+        pushEventHandlerProxy.shouldDisplayPushAppInForeground(push, completionHandler: { _ in
+            // When this block of code executes, the customer is done processing the push event.
 
-        // Call the completionHandler so customer does not need to. The push came from CIO, so it gets handled by the CIO SDK.
-        completionHandler(shouldShowPush)
+            // Because push came from CIO, ignore the return result of other push handlers.
+            // Determine if CIO push should be shown from SDK config
+            let shouldShowPush = self.moduleConfig.showPushAppInForeground
+
+            // The push came from CIO, so it gets handled by the CIO SDK.
+            // Calling the completion handler indicates to the OS that we are done processing the push.
+            completionHandler(shouldShowPush)
+        })
     }
 }